Check if Path Exists

+++ create-folder.js
@@ -2,10 +2,18 @@ const vscode = require("vscode");
 const fs = require("fs").promises;
 const path = require("path");
 
+async function pathExists(p) {
+	try {
+		await fs.access(p);
+		return true;
+	} catch {
+		return false;
+	}
+}
+
 async function createFolder(folder, window) {
 	try {
-		await fs.mkdir(folder, { recursive: true });
-		console.log(folder, " Created");
+		if (!pathExists(folder)) await fs.mkdir(folder, { recursive: true });
 	} catch (err) {
 		window.showErrorMessage(
 			`Failed to create folder ${folder}: ${err.message}`
@@ -15,8 +23,7 @@ async function createFolder(folder, window) {
 
 async function createFile(filePath, content = "", window) {
 	try {
-		await fs.writeFile(filePath, content);
-		console.log(filePath, " created");
+		if (!pathExists(filePath)) await fs.writeFile(filePath, content);
 	} catch (err) {
 		window.showErrorMessage(
 			`Failed to create file ${filePath}: ${err.message}`
